requirements.txt
# FastAPI web server
fastapi
uvicorn
python-dotenv
python-multipart

# Image processing
numpy
opencv-python-headless
Pillow
